fixing linters error
