+++ Gemfile
@@ -1,2 +1,5 @@
-gem 'rubocop', '>= 1.0', '< 2.0'
 source 'https://rubygems.org'
+# gem "rails"
+gem 'rspec', '~>3.0'
+
+gem 'rubocop', '>= 1.0', '< 2.0'
